+++ src/main/java/com/ydsw/domain/Sluice.java
@@ -6,6 +6,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 
@@ -66,7 +67,7 @@ public class Sluice implements Serializable {
      * 
      */
     @TableField(value = "angle")
-    private Integer angle;
+    private BigDecimal angle;
 
     /**
      *
